Update source code of bihype on ConvTran

+++ utils.py
@@ -149,8 +149,8 @@ class dataset_class(Dataset):
 
         y = self.labels[ind]  # (num_labels,) array
 
-        data = torch.tensor(x)
-        label = torch.tensor(y)
+        data = torch.tensor(x, dtype=torch.float32)
+        label = torch.tensor(y, dtype=torch.long)
 
         return data, label, ind
